Pseudoregalia: Double Pickups broke the autosplitter
If a player would hit the double pickup glitch, the splitter would before not see that the Item increased as it was specifically looking for an increase of 1 over a single frame.

Now we look for an increase of 1 to 3. We don't want just any increase as sometimes we can get randomly weird numbers for some items... though I think that is fixed now too?

Anyway: This now seems to work for anything I could throw at it under new understanding.